Вынесем общую логику в класс Smart
